calculateCP splitted
