+++ interface/src/App/PvE/Components/PveResult/PveResListFilter/PveResListSort/PveResListRender/PveResEntry/NameAndMoves/NameAndMoves.jsx
@@ -5,7 +5,8 @@ import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
 
 import WeatherMoves from "../WeatherMoves/WeatherMoves";
-import { calculateCP, calculateEffStat } from "js/indexFunctions";
+import { calculateEffStat } from "js/indexFunctions";
+import { calculateCP } from "js/cp/calculateCP";
 
 const NameAndMoves = React.memo(function NameAndMoves(props) {
     return (
